# openapi.json
{
  "components": {
    "schemas": {
      "AnnotationCreate": {
        "additionalProperties": false,
        "description": "One annotation to store, judged against the version its batch pinned.",
        "properties": {
          "asset_id": {
            "format": "uuid",
            "title": "Asset Id",
            "type": "string"
          },
          "attributes": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ]
            },
            "default": {},
            "title": "Attributes",
            "type": "object"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "geometry": {
            "discriminator": {
              "mapping": {
                "bbox": "#/components/schemas/BboxBody",
                "classification_tag": "#/components/schemas/ClassificationBody",
                "polygon": "#/components/schemas/PolygonBody"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/BboxBody"
              },
              {
                "$ref": "#/components/schemas/PolygonBody"
              },
              {
                "$ref": "#/components/schemas/ClassificationBody"
              }
            ],
            "title": "Geometry"
          },
          "label_class": {
            "title": "Label Class",
            "type": "string"
          },
          "model_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Ref"
          },
          "provenance": {
            "enum": [
              "human",
              "model",
              "import"
            ],
            "title": "Provenance",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "label_class",
          "geometry",
          "provenance"
        ],
        "title": "AnnotationCreate",
        "type": "object"
      },
      "AnnotationJobState": {
        "description": "Lifecycle: pending -> in_progress -> completed.\n\n``JOB_TRANSITIONS`` below is the whole of what is legal; ``JobService`` owns\nthe moves.",
        "enum": [
          "pending",
          "in_progress",
          "completed"
        ],
        "title": "AnnotationJobState",
        "type": "string"
      },
      "AnnotationOut": {
        "description": "One stored annotation, in the asset's own pixel frame.",
        "properties": {
          "asset_id": {
            "format": "uuid",
            "title": "Asset Id",
            "type": "string"
          },
          "attributes": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ]
            },
            "title": "Attributes",
            "type": "object"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "geometry": {
            "discriminator": {
              "mapping": {
                "bbox": "#/components/schemas/BboxBody",
                "classification_tag": "#/components/schemas/ClassificationBody",
                "polygon": "#/components/schemas/PolygonBody"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/BboxBody"
              },
              {
                "$ref": "#/components/schemas/PolygonBody"
              },
              {
                "$ref": "#/components/schemas/ClassificationBody"
              }
            ],
            "title": "Geometry"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label_class": {
            "title": "Label Class",
            "type": "string"
          },
          "model_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Ref"
          },
          "provenance": {
            "enum": [
              "human",
              "model",
              "import"
            ],
            "title": "Provenance",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "asset_id",
          "label_class",
          "schema_version",
          "geometry",
          "attributes",
          "provenance",
          "model_ref",
          "confidence"
        ],
        "title": "AnnotationOut",
        "type": "object"
      },
      "AnnotationPage": {
        "description": "A page of annotations.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnnotationOut"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "AnnotationPage",
        "type": "object"
      },
      "AnnotationUpdate": {
        "additionalProperties": false,
        "description": "One stored annotation, replaced whole.",
        "properties": {
          "attributes": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ]
            },
            "default": {},
            "title": "Attributes",
            "type": "object"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "geometry": {
            "discriminator": {
              "mapping": {
                "bbox": "#/components/schemas/BboxBody",
                "classification_tag": "#/components/schemas/ClassificationBody",
                "polygon": "#/components/schemas/PolygonBody"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/BboxBody"
              },
              {
                "$ref": "#/components/schemas/PolygonBody"
              },
              {
                "$ref": "#/components/schemas/ClassificationBody"
              }
            ],
            "title": "Geometry"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label_class": {
            "title": "Label Class",
            "type": "string"
          },
          "model_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Ref"
          },
          "provenance": {
            "enum": [
              "human",
              "model",
              "import"
            ],
            "title": "Provenance",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label_class",
          "geometry",
          "provenance"
        ],
        "title": "AnnotationUpdate",
        "type": "object"
      },
...
      "BboxBody": {
        "additionalProperties": false,
        "description": "An axis-aligned rectangle: top-left corner plus size, in asset pixels.",
        "properties": {
          "height": {
            "title": "Height",
            "type": "number"
          },
          "type": {
            "const": "bbox",
            "title": "Type",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "number"
          },
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "type",
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "BboxBody",
        "type": "object"
      },
...
      "ClassificationBody": {
        "additionalProperties": false,
        "description": "A whole-asset tag: a class with no coordinates.",
        "properties": {
          "type": {
            "const": "classification_tag",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "ClassificationBody",
        "type": "object"
      },
      "ErrorBody": {
        "description": "The one error shape this API emits, at every status.",
        "properties": {
          "code": {
            "description": "Stable machine-readable code. Branch on this, not on the status.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Extra structure whose shape depends on the code; absent when there is none.",
            "title": "Detail"
          },
          "message": {
            "description": "Human-readable sentence. Wording is not part of the contract.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "JobOut": {
        "description": "One annotator's unit of work over a segment of a batch.",
        "properties": {
          "asset_count": {
            "title": "Asset Count",
            "type": "integer"
          },
          "batch_id": {
            "format": "uuid",
            "title": "Batch Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/AnnotationJobState"
          }
        },
        "required": [
          "id",
          "batch_id",
          "state",
          "asset_count"
        ],
        "title": "JobOut",
        "type": "object"
      },
...
      "PolygonBody": {
        "additionalProperties": false,
        "description": "A closed polygon of at least three points. The closing edge is implicit.",
        "properties": {
          "points": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "title": "Points",
            "type": "array"
          },
          "type": {
            "const": "polygon",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "points"
        ],
        "title": "PolygonBody",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "A workspace API token, created with `visionset token create`. Sent as `Authorization: Bearer <token>`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST surface of the VisionSet SDK. The committed openapi.json is the contract.",
    "title": "Robomous VisionSet API",
    "version": "0.0.1.dev0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs/{job_id}": {
      "get": {
        "description": "The job, and the batch it is a segment of.\n\n`batch_id` is the handle worth having: it leads to the schema version this\njob's work is judged against, which a job id alone does not.",
        "operationId": "get_job",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request payload is not processable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unhandled server error, with an incident id"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The workspace is busy; retry after the header says"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/annotations": {
      "delete": {
        "description": "Remove annotations. One transaction, however many ids you pass.\n\nRepeating an id is not two deletions. An id that is not stored refuses the\nwhole call with 404 `ANNOTATION_NOT_FOUND` and removes nothing \u2014 there is no\npartial delete, for the reason there is no partial write.\n\nNo confirmation gate: taking a box off is the ordinary annotator edit loop,\nnot the destruction of a lifecycle entity. The batch gate is the guard, so\nonce the work closes nothing here can touch it at all.",
        "operationId": "delete_annotations",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "description": "An annotation to delete. Repeat the parameter per id.",
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "description": "An annotation to delete. Repeat the parameter per id.",
              "items": {
                "format": "uuid",
                "type": "string"
              },
              "title": "Id",
              "type": "array"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The resource's state refuses this request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request payload is not processable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unhandled server error, with an incident id"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The workspace is busy; retry after the header says"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Annotations",
        "tags": [
          "annotations"
        ]
      },
      "patch": {
        "description": "Replace stored annotations whole, judged against the same pinned version.\n\nAddressed by `id` and by nothing else \u2014 annotations are never reached by\nindex or position. There is no `asset_id` on the body because the stored one\nwins: moving a label from one asset to another is a delete and an add, not an\nedit, and doing it silently would take an asset's last annotation away\nwithout anything saying so.\n\nAll-or-nothing, and `detail.index` names the culprit, exactly as on the POST.",
        "operationId": "update_annotations",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/AnnotationUpdate"
                },
                "title": "Body",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The resource's state refuses this request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request payload is not processable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unhandled server error, with an incident id"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The workspace is busy; retry after the header says"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Annotations",
        "tags": [
          "annotations"
        ]
      },
      "post": {
        "description": "Store annotations, judged against the version this job's batch pinned.\n\nAll-or-nothing: every annotation is validated before any of them is written,\nso a payload with one bad box stores nothing at all. A half-labeled asset is\nnot a state a client can reach.\n\nA refusal that is about one item carries `detail.index` \u2014 the position in the\narray you sent \u2014 because nothing was written and the message alone cannot say\nwhich one it was. `schema_version` is not yours to set: the pinned version is\nstamped onto whatever you send, and comes back on the response.\n\nThe batch must be `in_annotation`, or this is 409\n`BATCH_NOT_IN_ANNOTATION`. An asset the job does not carry is 422\n`ASSET_NOT_IN_JOB`.",
        "operationId": "add_annotations",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/AnnotationCreate"
                },
                "title": "Body",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The resource's state refuses this request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request payload is not processable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unhandled server error, with an incident id"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The workspace is busy; retry after the header says"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Annotations",
        "tags": [
          "annotations"
        ]
      }
    },
    "/jobs/{job_id}/assets/{asset_id}/annotations": {
      "get": {
        "description": "Every annotation on one asset of this job, in the order they were added.\n\nEmpty for an asset nobody has labeled yet \u2014 the ordinary starting state, not\nan error. Reading is not gated on the batch being open: a label outlives the\nwork that produced it.",
        "operationId": "list_asset_annotations",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request payload is not processable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unhandled server error, with an incident id"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The workspace is busy; retry after the header says"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Asset Annotations",
        "tags": [
          "annotations"
        ]
      }
    },
...
  }
}